[Feature] player stats scraping
Added utility scraper which will pull all player stats data from
roster page and append data to SeasonScraper table as a single
column of dictionaries.

Every entry in the column is a nested dictionary of the following
format:

{ "Table 1":
	{ "col 1": ,
	  "col 2": ,
		...
	} ,
  "Table 2":
	{ "col 1": ,
	  "col 2": ,
		...
	},
	...
}

This data will need to be unpacked to actually be usable in a model.